requirements.txt
redis
gunicorn
wheel
pillow
beautifulsoup4==4.10.0
celery==5.1.2
django-celery-beat==2.2.1
Django==3.2.8
djangorestframework==3.12.4
pytz==2021.3
requests==2.26.0
